Update us_atrocities.md, add paragraph about the Huk rebellion in Philippines
This PR includes a new paragraph about the US war against the Huks in Philippines between 1945-54.

Some of the information is mentioned in the Wikipedia page but most of the info is from a book: Shalom, Stephen. (1986) THE UNITED STATES AND THE PHILIPPINES: A STUDY OF NEOCOLONIALISM. Quezon City: New Day Publishers. There is a direct PDF link to the book at the end of the paragraph and the PDF is searchable with using Ctrl+F.